% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asl_func.R
\name{ASL_boilerplate}
\alias{ASL_boilerplate}
\title{ASL Boilerplate}
\usage{
ASL_boilerplate(
  stratified = NULL,
  abundance = NULL,
  data = NULL,
  species = "fish",
  tense = c(NA, "past", "present", "future"),
  age = NULL,
  sex = NULL,
  length = NULL,
  stratum = NULL,
  Nhat = NULL,
  se_Nhat = NULL,
  stratum_weights = NULL,
  verbose = FALSE,
  FPC = c("ifknown", "always", "never")
)
}
\arguments{
\item{stratified}{Whether stratified estimators will be used (\code{TRUE}) or not (\code{FALSE})}

\item{abundance}{Possible values of \code{c("known", "estimated", "unknown")},
defining how abundance is treated.}

\item{data}{A vector containing some subset of \code{c("age", "sex", "length")},
depending on what data are present}

\item{species}{Optional character to use for species.  Defaults to \code{"fish"}.}

\item{tense}{Which verb tense to use.  If the default \code{NA} is accepted, this
will be imputed as \code{"future"} if text is constructed from method 1 above,
or \code{"past"} if text is constructed from data (method 2 above).  Possible
values are \code{c(NA, "past", "present", "future")}}

\item{age}{Argument to \code{ASL_table()}.  Defaults to \code{NULL}.}

\item{sex}{Argument to \code{ASL_table()}.  Defaults to \code{NULL}.}

\item{length}{Argument to \code{ASL_table()}.  Defaults to \code{NULL}.}

\item{stratum}{Argument to \code{ASL_table()}.  Defaults to \code{NULL}.}

\item{Nhat}{Argument to \code{ASL_table()}.  Defaults to \code{NULL}.}

\item{se_Nhat}{Argument to \code{ASL_table()}.  Defaults to \code{NULL}.}

\item{stratum_weights}{Argument to \code{ASL_table()}.  Not used.}

\item{verbose}{Argument to \code{ASL_table()}.  Not used.}

\item{FPC}{Argument to \code{ASL_table()}.  Possible arguments are \code{c("ifknown", "always", "never")}.  Defaults to \code{"ifknown"}.}
}
\description{
Automagically creates boilerplate text to accompany an ASL
summary analysis, or methods for an Operational Plan.

This can be done two ways:
\itemize{
\item (1) Directly from arguments \verb{stratified=}, \verb{abundance=}, and \verb{data=} (see argument description below)
\item (2) From possible inputs to \link{ASL_table}.  This method is provided for convenience.
}

Text and equations are taken directly from the document found here:

\verb{https://github.com/ADFG-DSF/dsftools/blob/main/addl_documentation/asl_equations.Rmd}
}
\examples{
\dontrun{

## An example Rmarkdown code chunk might look like the following:

```{r, results='asis', echo=FALSE}
library(dsftools)
ASL_boilerplate(data=c("age", "length"), stratified=TRUE, abundance="known")
```
}
}
\seealso{
\link{ASL_table}
}
\author{
Matt Tyers
}
