% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asl_func.R
\name{ASL_table}
\alias{ASL_table}
\title{Construct an ASL Table from Vectors of Data}
\usage{
ASL_table(
  age = NULL,
  sex = NULL,
  length = NULL,
  stratum = NULL,
  Nhat = NULL,
  se_Nhat = NULL,
  stratum_weights = NULL,
  verbose = FALSE,
  FPC = c("ifknown", "always", "never")
)
}
\arguments{
\item{age}{Vector of ages.  If the default (\code{NULL}) is accepted, age will be
omitted from the resulting table.}

\item{sex}{Vector of sex.  If the default (\code{NULL}) is accepted, sex will be
omitted from the resulting table.}

\item{length}{Vector of length.  If the default (\code{NULL}) is accepted, length
will be omitted from the resulting table.}

\item{stratum}{Optional vector of stratum, if stratified estimators will be
used.  This must be formatted as positive whole numbers (1, 2, 3, etc).
Defaults to \code{NULL}.}

\item{Nhat}{Optional vector of abundance for each stratum, if stratified
estimators will be used.  The length of this vector must correspond to the
maximum value in the \verb{stratum=} argument.  Defaults to \code{NULL}.}

\item{se_Nhat}{Optional vector of the standard error of abundance for each
stratum, if stratified estimators will be used and by-stratum abundance is
considered to have measurement error.    The length of this vector must
correspond to the maximum value in the \verb{stratum=} argument.  Defaults to \code{NULL}.}

\item{stratum_weights}{Optional vector of weights for each stratum, if
relative weights are known but abundance is not.  Using this argument rather
than \code{Nhat} will also omit the Finite Population Correction factor (FPC) from
calculations, but weights are treated as constant (without error), likely
underestimating variance.  Defaults to \code{NULL}.}

\item{verbose}{Whether to print messages corresponding to the method used.
Defaults to \code{FALSE}.}

\item{FPC}{Whether to incorporate the Finite Population Correction factor (FPC) in
variance calculations.  Allowed values are
\itemize{
\item \code{"ifknown"} (the default), which will use the FPC only if abundance is known without error
\item \code{"always"}, which will use the FPC wherever possible (i.e. if there is an estimate of abundance)
\item \code{"never"}, which will always ignore the FPC.
used if the abundance is considered to be known without error.
}}
}
\value{
A data.frame with rows corresponding to categories of age and/or sex,
depending on data inputs.Stratified - If abundance is known without error
\subsection{If there are proportions}{

The proportion of each age and/or sex category \emph{z} will be estimated for each sampling stratum \emph{t} as follows:

\deqn{\hat{p}_{tz}=\frac{n_{tz}}{n_t}}

in which \eqn{n_{tz}} equals the number of fish sampled during sampling stratum \eqn{t} classified as age and/or sex category \eqn{z}, and \eqn{n_t} equals the number of fish sampled for age and/or sex determination within sampling stratum \eqn{t}.

The sampling variance of \eqn{\hat{p}_{tz}} will be estimated as the following (Cochran 1977):

\deqn{\hat{var}[\hat{p}_{tz}]=\frac{\hat{p}_{tz}(1-\hat{p}_{tz})}{n_t-1}\left(\frac{N_t-n_t}{N_t-1}\right)}

if \eqn{N_t}, the total abundance of fish in sampling stratum \eqn{t}, is known and the finite population correction factor (FPC) is used; otherwise, as the following:

\deqn{\hat{var}[\hat{p}_{tz}]=\frac{\hat{p}_{tz}(1-\hat{p}_{tz})}{n_t-1}}

The total abundance by age and/or sex category in each sampling stratum will be estimated as follows:

\deqn{\hat{N}_{tz}=N_t\hat{p}_{tz}}

with variance estimated as

\deqn{\hat{var}[\hat{N}_{tz}]=N_t^2\hat{var}[\hat{p}_{tz}]}

The total abundance by age and/or sex category and its variance will then be estimated by summation as follows:

\deqn{\hat{N}_z=\sum_{t=1}^{L}\hat{N}_{tz}}

and

\deqn{\hat{var}[\hat{N}_{z}]=\sum_{t=1}^{L}\hat{var}[\hat{N}_{tz}]}

where \eqn{L} equals the number of sampling strata.

Finally, the overall proportion by age and/or sex category and its variance will be estimated as follows:

\deqn{\hat{p}_z=\frac{\hat{N}_z}{N}}

and

\deqn{\hat{var}[\hat{p}_z]=\frac{\hat{var}[\hat{N}_z]}{N^2}}

where \eqn{N} is the total abundance across all sampling periods.

The mean length by age and/or sex for each sampling stratum will be estimated as follows:

\deqn{\bar{x}_{tz}=\frac{\sum_{i=1}^{n_{tz}}x_{tzi}}{n_{tz}}}

where \eqn{x_{tzi}} is the length of the \emph{i}th fish sampled of age and/or sex \eqn{z} during sampling stratum \eqn{t}.

The sampling variance of \eqn{\bar{x}_{tz}} will be estimated as

\deqn{\hat{var}[\bar{x}_{tz}]=\frac{\sum_{i=1}^{n_{tz}}(x_{tzi}-\bar{x}_{tz})^2}{n_{tz}(n_{tz}-1)}\left(\frac{
    \hat{N}_{tz}-n_{tz}}{\hat{N}_{tz}-1}\right)}

if the finite population correction factor (FPC) will be used; otherwise, as the following:

\deqn{\hat{var}[\bar{x}_{tz}]=\frac{\sum_{i=1}^{n_{tz}}(x_{tzi}-\bar{x}_{tz})^2}{n_{tz}(n_{tz}-1)}}

The mean length by age and/or sex category will then be estimated as follows:

\deqn{\bar{x}_z=\sum_{t=1}^{L}\frac{\hat{N}_{tz}}{\hat{N}_z}\bar{x}_{tz}}

with its variance approximated using a Taylor's series expansion (Mood et al. 1974):

\deqn{\hat{var}[\bar{x}_z]\approx\sum_{t=1}^{L}\frac{\hat{N}_{tz}^2}{\hat{N}_z^2}\hat{var}[\bar{x}_{tz}]+\sum_{t=1}^{L}\frac{\left(\bar{x}_{tz}\hat{N}_z-\left(\sum_{u=1}^{L}\bar{x}_{uz}\hat{N}_{uz}\right)\right)^2}{\hat{N}_z^4}\hat{var}[\hat{N}_{tz}]}
}

\subsection{If there are no proportions to estimate}{

The mean length for each sampling stratum will be estimated as follows, where \eqn{x_{ti}} is the length of the \emph{i}th fish sampled within sampling stratum \eqn{t}, and \eqn{n_t} is the number of fish in stratum \emph{t} sampled for length:

\deqn{\bar{x}_{t}=\frac{\sum_{i=1}^{n_{t}}x_{ti}}{n_{t}}}

The sampling variance of \eqn{\bar{x}_{t}} will be estimated as

\deqn{\hat{var}[\bar{x}_{t}]=\frac{\sum_{i=1}^{n_{t}}(x_{ti}-\bar{x}_{t})^2}{n_{t}(n_{t}-1)}\left(\frac{
N_{t}-n_{t}}{N_{t}-1}\right)}

if abundance per stratum \eqn{N_t} is known and if the finite population correction factor is used, otherwise as:

\deqn{\hat{var}[\bar{x}_{t}]=\frac{\sum_{i=1}^{n_{t}}(x_{ti}-\bar{x}_{t})^2}{n_{t}(n_{t}-1)}}

Stratified estimates of mean length will be calculated as follows, in which \eqn{N_t} and \eqn{\bar{x}_t} represent the abundance and average length associated with sampling stratum \eqn{t}, respectively, and \eqn{N} represents the total abundance:

\deqn{\bar{x}=\frac{1}{N}\sum_{t=1}^L N_t\bar{x}_t}

and

\deqn{\hat{var}[\bar{x}]=\sum_{t=1}^L\left(\frac{N_t}{N}\right)^2 \hat{var}[\bar{x}_t]}
}
}
\description{
This function is intended to perform all the necessary calculations
to construct tables summarizing ASL (Age, Sex, and Length) for several data
scenarios:
\itemize{
\item Stratified vs. non-stratified
\item Abundance known without error vs. estimated with error vs. unknown
\item Data available for Age, Sex, and Length vs. some subset.
}

The function will return a table with rows for each combination of Age and
Sex (as available), and will summarize the respective proportions, estimated
abundance, and associated Lengths (as available).

Generally, \code{NULL} values in a given data vector indicate that the vector will
not be used.
}
\examples{
### a look at input
head(sim_data$data)
sim_data$abundance

### tables for possible scenarios

# Stratified with error in Nhat
ASL_table(age=sim_data$data$age,
          length=sim_data$data$length,
          stratum=sim_data$data$stratum,
          Nhat=sim_data$abundance$Nhat,
          se_Nhat=sim_data$abundance$se_Nhat)

# Stratified without error in Nhat
ASL_table(age=sim_data$data$age,
          length=sim_data$data$length,
          stratum=sim_data$data$stratum,
          Nhat=sim_data$abundance$Nhat)

# Pooled (not stratified) with error in Nhat
ASL_table(age=sim_data$data$age,
          length=sim_data$data$length,
          Nhat=100*1000,
          se_Nhat=10*1000)

# Pooled (not stratified) without error in Nhat
ASL_table(age=sim_data$data$age,
          length=sim_data$data$length,
          Nhat=100*1000)
}
\seealso{
\link{verify_ASL_table}
}
\author{
Matt Tyers
}
