% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{binomial_detection}
\alias{binomial_detection}
\title{Detection Probability From a Given Sample Size, Assuming Binomial (Deprecated)}
\usage{
binomial_detection(
  n_raw,
  prop_usedby,
  assumed_survival = 1,
  observe_at_least = 1
)
}
\arguments{
\item{n_raw}{Trial value of sample size, before accounting for mortality.}

\item{prop_usedby}{Hypothetical proportion of the population using the area
considered.}

\item{assumed_survival}{Assumed survival (or 1 - data loss proportion).
Defaults to \code{1}.}

\item{observe_at_least}{Minimum number of marked individuals to consider as
detection of an area.  Defaults to \code{1}, but a larger number may be used as
necessary, depending on criterion used to define detection of an aggregation.}
}
\value{
A single number reflecting the probability of detection.
}
\description{
In the context of a telemetry study, this function calculates
the probability of detecting a SINGLE area used by some proportion of the marked
population, given the sample size of instrumented fish, assuming a Binomial
distribution with random sampling.

This is actually a simple wrapper of the function call below, and re-written
for convenience:

\verb{pbinom(q = observe_at_least-1,}

\verb{       size = round(n_raw*assumed_survival),}

\verb{       prob = prop_usedby,}

\verb{       lower.tail = FALSE)}

Because of this, vectors may be used in arguments, but care should be taken
to ensure the function evaluates as desired (see examples below).

If the probability of simultaneous detection among multiple areas is desired,
\link{multinomial_detection} may be considered.
}
\examples{
## The probability of detecting an area used by 5\% of the population, given
## a sample size of n=100 and assuming 80\% survival.
suppressWarnings({
binomial_detection(n_raw=100, prop_usedby = 0.05, assumed_survival = .8)
})

## examples with vector-valued input:
suppressWarnings({
binomial_detection(n_raw=c(80, 100), prop_usedby = 0.05, assumed_survival = .8)
})
suppressWarnings({
binomial_detection(n_raw=100, prop_usedby = c(0.025, 0.05), assumed_survival = .8)
})

## if multiple vectors are used, evaluation will be strictly vector-wise
suppressWarnings({
binomial_detection(n_raw=c(80, 100), prop_usedby = c(0.025, 0.05))
})

## but outer() may be used if a 2d matrix is desired?
n_raw_trial <- c(80, 100)
prop_usedby_trial <- c(0.025, 0.05)
names(n_raw_trial) <- n_raw_trial              ## optional but useful
names(prop_usedby_trial) <- prop_usedby_trial  ## optional but useful
suppressWarnings({
outer(n_raw_trial, prop_usedby_trial, FUN=binomial_detection)
})
suppressWarnings({
outer(n_raw_trial, prop_usedby_trial, FUN=binomial_detection, assumed_survival=0.8)
})

## using outer() with different arguments
n_raw_trial <- c(80, 100)
survival_trial <- c(1, 0.9, 0.8)
suppressWarnings({
outer(X=n_raw_trial, Y=survival_trial,
      FUN=function(X,Y) binomial_detection(n_raw=X,
                                           prop_usedby=0.05,
                                           assumed_survival=Y))
})
}
\seealso{
\link{detection_probability}
}
\author{
Matt Tyers
}
