% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{expit}
\alias{expit}
\title{Expit, or inverse logit}
\usage{
expit(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Inverse logit, where logit is defined as \eqn{log\left(\frac{x}{1-x}\right)}.

Expit (inverse logit) is defined as \eqn{\frac{exp(x)}{1+exp(x)}}
}
\examples{
expit(0)

x <- seq(from=-5, to=5, by=0.1)
plot(x, expit(x))
}
\seealso{
\link{logit}
}
\author{
Matt Tyers
}
