% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{logit}
\alias{logit}
\title{Logit}
\usage{
logit(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Logit, defined as
\deqn{log\left(\frac{x}{1-x}\right)}
}
\examples{
logit(0.5)

x <- seq(from=0.01, to=0.99, by=0.01)
plot(x, logit(x))
}
\seealso{
\link{expit}
}
\author{
Matt Tyers
}
