% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{multinomial_detection}
\alias{multinomial_detection}
\title{Simultaneous Detection Probability From a Given Sample Size, Assuming Multinomial (Deprecated)}
\usage{
multinomial_detection(
  n_raw,
  prop_usedby,
  assumed_survival = 1,
  observe_at_least = 1,
  prop_ofareas = 1
)
}
\arguments{
\item{n_raw}{Trial value of sample size, before accounting for mortality.}

\item{prop_usedby}{Hypothetical proportion of the population using the area
considered.}

\item{assumed_survival}{Assumed survival (or 1 - data loss proportion).
Defaults to \code{1}.}

\item{observe_at_least}{Minimum number of marked individuals to consider as
detection of an area.  Defaults to \code{1}, but a larger number may be used as
necessary, depending on criterion used to define detection of an aggregation.}

\item{prop_ofareas}{Proportion of areas to simultaneously detect.  It may be
desirable to structure a precision statement in terms of detection of some
percentage of areas, see examples below.  Defaults to \code{1}, which can be
interpreted as simultaneous detection of all such areas.}
}
\value{
Named list of estimated probabilities:
\itemize{
\item Element \verb{$avg_p_detected} gives the average estimated detection probability
over all simulated areas, which will be a good estimate of the detection
probability of a SINGLE given area.
\item Element \verb{$p_all_detected} gives the estimated SIMULTANEOUS detection probability
of the proportion of areas specified.
}
}
\description{
In the context of a telemetry study, this function uses simulation
to estimate the probability of detecting a single area used by some proportion
of the marked population, given the sample size of instrumented fish, as well
as the probability of detecting MULTIPLE areas, assuming random sampling from
a Multinomial worst-case scenario.

The multinomial worst-case scenario is defined as a distribution with equal
probabilities of the proportion specified by \verb{prop_usedby=}.
For example, if the proportion supplied is 5\%, a multinomial distribution with
twenty (=1/0.05) categories with equal probabilities will be simulated.

If the probability of detection of just a single area is desired,
\link{binomial_detection} may be considered as an exact solution.
}
\examples{
## The probability of detecting all areas used by 5\% of the population, given
## a sample size of n=80 and assuming 80\% survival, is at least 44\%.
multinomial_detection(n_raw = 80, prop_usedby = 0.05, assumed_survival = .8)


## The probability of detecting 90\% of areas used by 5\% of the population, given
## a sample size of n=80 and assuming 80\% survival, is at least 97\%.
multinomial_detection(n_raw = 80, prop_usedby = 0.05, assumed_survival = .8,
                      prop_ofareas = 0.9)
}
\seealso{
\link{binomial_detection}
}
\author{
Matt Tyers
}
