% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{rp}
\alias{rp}
\title{Relative Precision from Desired Confidence or Accuracy}
\usage{
rp(
  sim_vec,
  true_val,
  confidence = c(0.8, 0.9, 0.95, 0.99),
  accuracy = NA,
  relative = TRUE
)
}
\arguments{
\item{sim_vec}{Numeric vector of simulations}

\item{true_val}{Assumed true value, which may be given as a single number or
corresponding vector if needed}

\item{confidence}{Vector of desired confidence levels.  Defaults to
\code{c(0.8, 0.9, 0.95, 0.99)}.}

\item{accuracy}{Vector of desired (relative) accuracy.  Defaults to \code{NA}, which
will calculate (relative) accuracy from precision.  If this argument is given a
numeric input, the function will calculate precision from (relative) accuracy.}

\item{relative}{Whether to calculate (or calculate from) RELATIVE accuracy or
ABSOLUTE accuracy.  Defaults to \code{TRUE}, which will specify relative accuracy.}
}
\value{
Named vector of values of (relative) accuracy or confidence, depending
on usage.
}
\description{
Calculates (relative) accuracy from desired confidence, or
confidence from desired (relative) accuracy, from a vector of simulated values
and the true value.

This function is intended to fill in the typical language in an Operational
Plan, stating "...estimate (some parameter), such that the estimate is within
(\code{accuracy}) of the true value (\code{confidence}) percent of the time."

This function can calculate (relative) accuracy from desired confidence, or
confidence from desired (relative) accuracy, depending on whether the \verb{accuracy=}
argument is specified.  See examples below for both usage cases.

Depending on the use case, it may be more appropriate to calculate (or
calculate from) relative accuracy, which is defined as
\eqn{|\frac{estimate - truth}{truth}|}
or absolute accuracy, which is defined as
\eqn{|estimate - truth|}.
}
\examples{
## first simulating a random vector
xx <- rnorm(10000, mean=10, sd=1)

## calculating RELATIVE accuracy from desired confidence
rp(xx, true_val=10)

## calculating confidence from desired RELATIVE accuracy
rp(xx, true_val=10, accuracy=seq(from=0.1, to=0.5, by=0.1))

## calculating ABSOLUTE accuracy from desired confidence
rp(xx, true_val=10, relative=FALSE)

## calculating confidence from desired ABSOLUTE accuracy
rp(xx, true_val=10, accuracy=seq(from=0.5, to=2.5, by=0.5), relative=FALSE)

## plotting both methods and showing agreement
plot(x = seq(.01, .5, by=.01),
     y = rp(xx, true_val=10, accuracy=seq(.01, .5, by=.01)),
     xlab = "relative accuracy", ylab="confidence")
points(x = rp(xx, true_val=10, confidence=c(0.8, 0.9, 0.95, 0.99)),
       y = c(0.8, 0.9, 0.95, 0.99), pch=16, col=2)
}
\author{
Matt Tyers
}
