% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{se}
\alias{se}
\title{Standard Error}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{Logical.  Should missing values be removed?  Defaults to \code{FALSE}
for consistency with \link{sd}.  Note: if \code{na.rm==TRUE}, the denominator will
be the number of non-NA entries.}
}
\value{
Numeric of length 1
}
\description{
Standard error, defined as
\eqn{\frac{\sigma_x}{\sqrt{n}}}
}
\examples{
a <- c(8, 6, 7, 5, 3, 0, 9, NA)

se(a)
se(a, na.rm=TRUE)
sd(a, na.rm=TRUE)/sqrt(7)
}
\author{
Matt Tyers
}
