% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asl_func.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Example data: sim_data}
\format{
An object of class \code{list} of length 2.
}
\usage{
sim_data
}
\description{
A simulated dataset intended to illustrate \link{ASL_table}.  This dataset
is formatted as a list with two components:
\itemize{
\item \code{sim_data$data} is a data frame with 400 observations of \verb{$age}, \verb{$length},
and \verb{$stratum}.
\item \code{sim_data$abundance} is a data frame with 4 rows of \verb{$Nhat} and \verb{$se_Nhat},
corresponding to each stratum.
}
}
\keyword{datasets}
