% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asl_sim.R
\name{simulate_ASL_table}
\alias{simulate_ASL_table}
\title{Simulation of many replicates of \code{ASL_table()}}
\usage{
simulate_ASL_table(
  case = NULL,
  nstrata,
  nage,
  nt,
  Nt,
  stratum_weights = NULL,
  se_Nt = NULL,
  mn_length = NULL,
  sd_length = NULL,
  ptz = NULL,
  nNA = 0,
  nsim = 1000,
  plot_pop = TRUE,
  verbose = TRUE,
  print_table = FALSE
)
}
\arguments{
\item{case}{If a pre-determined case is to be used.  Allowed values are
\code{"stratified_witherror_lengthage"}, \code{"stratified_witherror_age"}, \code{"stratified_witherror_length"},
\code{"stratified_lengthage"}, \code{"stratified_age"}, \code{"stratified_length"},
\code{"stratified_Nunknown_lengthage"}, \code{"stratified_Nunknown_age"}, \code{"stratified_Nunknown_length"},
\code{"pooled_witherror_lengthage"}, \code{"pooled_witherror_age"}, \code{"pooled_witherror_length"}
\code{"pooled_lengthage"}, \code{"pooled_age"},  \code{"pooled_length"},
\code{"pooled_Nunknown_lengthage"}, \code{"pooled_Nunknown_length"}, or \code{"pooled_Nunknown_age"}. If the default
(\code{NULL}) is accepted, all simulation parameters below must be supplied by the user.}

\item{nstrata}{Number of sampling strata}

\item{nage}{Number of age categories}

\item{nt}{Sample size for each stratum}

\item{Nt}{Abundance for each stratum.  If abundance is completely unknown, sampling
weights may be used instead in the \verb{sampling_weights=} argument.}

\item{stratum_weights}{Optional vector of sampling weights for each stratum.  Defaults to \code{NULL}.}

\item{se_Nt}{Optional vector of standard errors of abundance for each stratum
Defaults to \code{NULL}, implying abundance is known without error.}

\item{mn_length}{Vector of mean lengths for each age category.  Defaults to
\code{NULL}, implying lengths were not considered.}

\item{sd_length}{Vector of standard deviations for each age category.  Defaults to
\code{NULL}, implying lengths were not considered.}

\item{ptz}{Matrix of probabilities of each age by stratum, with rows
corresponding to strata and columns corresponding to ages.  The probabilities
for each row will be normalized before simulation, therefore summing to one
is not required.  If a pooled (non-stratified) sample is to be taken and
age categories are to be considered, this should be supplied as a vector with
length equal to the number of ages.}

\item{nNA}{Number of NA values to randomly impute, to test robustness to NA.  Defaults to \code{0}.}

\item{nsim}{Number of simulated replicates.  Defaults to \code{1000}, but more is recommended.}

\item{plot_pop}{Whether to make summary plots of the simulated population and
one representative sample, in addition to the plots produced in simulation.
Defaults to \code{TRUE}.}

\item{verbose}{Whether to print the parameters used in simulation to the console,
if one of the \code{case}s is accepted, as well as printing the method used within
\code{ASL_table()}.  Defaults to \code{TRUE}.}

\item{print_table}{Whether to print an example output table from \code{ASL_table()}
as an additional check.  Defaults to \code{FALSE}.}
}
\value{
An object of class \code{"ASL_sim"}, which is a list of matrices of
simulation samples, and vectors of the associated
true values.  Details are described above.
}
\description{
This function is called internally within
\link{verify_ASL_table} and \link{rp_ASL_table}.  However, it may be advantageous
to first create simulation results if there will be multiple calls to
\code{verify_ASL_table()} or \code{rp_ASL_table()}.

Eighteen pre-determined cases may be used in the \verb{case=} argument, depending on
whether a stratified sampling scheme is used, whether abundance is known,
estimated with error, or completely unknown, and whether estimates pertaining to categorical (age)
and/or numeric (length) values is to be performed.  In general:
\itemize{
\item Five age classes are simulated (if age is considered)
\item Mean and standard deviation of length increase with age (if length and age are considered)
\item Data are treated as belonging to four temporal strata (if stratified), in which:
\itemize{
\item Age class generally decreases with stratum (older fish arrive sooner)
\item Abundance and its standard error both increase with stratum (larger numbers of fish arrive later), but
\item An equal number of fish are sampled per temporal stratum.
}
}

It should be noted that a small but non-trival number of entries for age,
length, and stratum are imputed as \code{NA} by default (see \verb{nNA=} argument) in order
to test function robustness to missing data.

A list will be returned, with matrices of simulation samples and associated
true values.  This will consist of some subset of the following, as appropriate:
\itemize{
\item Proportions
\itemize{
\item \verb{$phat_sim}: A matrix of simulated proportion estimates, with columns corresponding
to categories and rows corresponding to simulation samples.  This is calculated
via the relevant methods used by \code{ASL_table()}.
\item \verb{$phat_true}: A vector of true proportions.  This is calculated from simulation
inputs \code{Nt} or \code{ptz} and \code{stratum_weights} as appropriate.
\item \verb{$se_phat_sim}: A matrix of simulated proportion standard errors, with columns corresponding
to categories and rows corresponding to simulation samples.  This is calculated
via the relevant methods used by \code{ASL_table()}.
\item \verb{$se_phat_true}: A vector of standard errors, calculated empirically from
the simulated values \verb{$phat_sim} thereby approximating the sampling distribution.
}
\item Abundances
\itemize{
\item \verb{$Nhat_sim}: A matrix of simulated abundance estimates, with columns corresponding
to categories and rows corresponding to simulation samples.  This is calculated
via the relevant methods used by \code{ASL_table()}.
\item \verb{$Nhat_true}: A vector of true abundances, given by user inputs \code{Nt}.
\item \verb{$se_Nhat_sim}: A matrix of simulated abundance standard errors, with columns corresponding
to categories and rows corresponding to simulation samples.  This is calculated
via the relevant methods used by \code{ASL_table()}.
\item \verb{$se_Nhat_true}: A vector of standard errors, calculated empirically from
the simulated values \verb{$Nhat_sim} thereby approximating the sampling distribution.
}
\item Mean Lengths
\itemize{
\item \verb{$mn_length_sim}: A matrix of simulated mean length estimates, with columns corresponding
to categories and rows corresponding to simulation samples.  This is calculated
via the relevant methods used by \code{ASL_table()}.
\item \verb{$mn_length_true}: A vector of true mean lengths, given by user inputs \code{mn_length}.
\item \verb{$se_mn_length_sim}: A matrix of simulated mean length standard errors, with columns corresponding
to categories and rows corresponding to simulation samples.  This is calculated
via the relevant methods used by \code{ASL_table()}.
\item \verb{$se_mn_length_true}: A vector of standard errors, calculated empirically from
the simulated values \verb{$mn_length_sim} thereby approximating the sampling distribution.
}
}
}
\examples{
## creating a simulation object using case=
simresults <- simulate_ASL_table(case="stratified_witherror_lengthage",
                                 nsim=500, plot_pop=FALSE)

## creating a simulation object using simulation parameters
simresults <- simulate_ASL_table(nstrata = 4,
                                 nage = 5,

                                 # sample size for each stratum
                                 nt = c(100, 100, 100, 100),

                                 # abundance for each stratum
                                 Nt = c(10000, 20000, 30000, 40000),

                                 # (possible) SE for abundance by stratum
                                 se_Nt = 0.2*c(10000, 20000, 30000, 40000),

                                 # mean length FOR EACH AGE
                                 mn_length = c(150, 200, 250, 300, 350),

                                 # sd length FOR EACH AGE
                                 sd_length = c(40, 50, 60, 70, 80),

                                 # matrix of probabilities of each age BY stratum
                                 ptz = matrix(c(c(1,2,3,4,5),
                                                c(1,2,5,5,2),
                                                c(2,5,3,2,1),
                                                c(5,4,3,1,1)), byrow=TRUE, nrow=4, ncol=5),
                                 nsim=500, plot_pop=FALSE)

## running rp_ with the object created
par(mfrow=c(2,2))
rp_ASL_table(sim=simresults)

## running rp_ again
par(mfrow=c(2,2))
rp_ASL_table(sim=simresults, conf_target=0.99)

## running verify_
par(mfrow=c(3,2))
verify_ASL_table(sim=simresults)
}
\seealso{
\link{rp_ASL_table}, \link{verify_ASL_table}, \link{ASL_table}
}
\author{
Matt Tyers
}
