% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_subset.R
\name{subsets}
\alias{subsets}
\alias{\%s_l\%}
\alias{\%s_g\%}
\alias{\%s_leq\%}
\alias{\%s_geq\%}
\alias{\%s_inside\%}
\alias{\%s_inside()\%}
\alias{\%s_inside(]\%}
\alias{\%s_inside[)\%}
\title{Vector subsetting shorthand}
\usage{
x \%s_l\% a

x \%s_g\% a

x \%s_leq\% a

x \%s_geq\% a

x \%s_inside\% a

x \%s_inside()\% a

x \%s_inside(]\% a

x \%s_inside[)\% a
}
\arguments{
\item{x}{Vector input}

\item{a}{Number or vector for logical comparison}
}
\value{
Subset of vector \code{x}
}
\description{
Shortcut functions returning a subset of some vector.  These
may be handy when dealing with vectors or data.frame columns with long names.

\verb{\%s_l\%}, \verb{\%s_leq\%}, \verb{\%s_g\%}, and \verb{\%s_geq\%} denote the subset less than, less
than or equal to, greater than, and greater than or equal to some quantity.

\code{x \%s_l\% a} is evaluated simply as \code{x[x < a]}, preserving all vector-wise
behavior if it is desirable to exploit this.

\verb{\%s_inside\%} and friends denote the subset bounded by the range expressed by
some vector (see \link{inside}).
}
\examples{
## subsets in comparison to a single number
1:10 \%s_l\% 5
1:10 \%s_leq\% 5
1:10 \%s_g\% 5
1:10 \%s_geq\% 5

## subsets in comparison to a range
1:10 \%s_inside\% 3:7
1:10 \%s_inside()\% 3:7
1:10 \%s_inside[)\% 3:7
1:10 \%s_inside(]\% 3:7
}
\author{
Matt Tyers
}
