% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asl_func.R
\name{verify_ASL_table}
\alias{verify_ASL_table}
\title{Empirical Verification of the Methods Used in \code{ASL_table()}}
\usage{
verify_ASL_table(
  case = NULL,
  nstrata,
  nage,
  nt,
  Nt,
  se_Nt = NULL,
  mn_length = NULL,
  sd_length = NULL,
  ptz = NULL,
  nsim = 1000,
  plot_pop = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{case}{If a pre-determined case is to be used.  Allowed values are
\code{"stratified_witherror_lengthage"}, \code{"stratified_witherror_age"}, \code{"stratified_witherror_length"},
\code{"stratified_lengthage"}, \code{"stratified_age"}, \code{"stratified_length"},
\code{"pooled_witherror_lengthage"}, \code{"pooled_witherror_age"}, \code{"pooled_witherror_length"}
\code{"pooled_lengthage"}, \code{"pooled_age"}, or \code{"pooled_length"}. If the default
(\code{NULL}) is accepted, all simulation parameters below must be supplied by the user.}

\item{nstrata}{Number of sampling strata}

\item{nage}{Number of age categories}

\item{nt}{Sample size for each stratum}

\item{Nt}{Abundance for each stratum}

\item{se_Nt}{Optional vector of standard errors of abundance for each stratum
Defaults to \code{NULL}, implying abundance is known without error.}

\item{mn_length}{Vector of mean lengths for each age category.  Defaults to
\code{NULL}, implying lengths were not considered.}

\item{sd_length}{Vector of standard deviations for each age category.  Defaults to
\code{NULL}, implying lengths were not considered.}

\item{ptz}{Matrix of probabilities of each age by stratum, with rows
corresponding to strata and columns corresponding to ages.  The probabilities
for each row will be normalized before simulation, therefore summing to one
is not required.  If a pooled (non-stratified) sample is to be taken and
age categories are to be considered, this should be supplied as a vector with
length equal to the number of ages.}

\item{nsim}{Number of simulated replicates.  Defaults to \code{1000}, but more is recommended.}

\item{plot_pop}{Whether to make summary plots of the simulated population and
one representative sample, in addition to the plots produced in simulation.
Defaults to \code{TRUE}.}

\item{verbose}{Whether to print the parameters used in simulation to the console,
if one of the \code{case}s is accepted, as well as printing the method used within
\code{ASL_table()}.  Defaults to \code{TRUE}.}
}
\value{
\code{NULL}
}
\description{
This function draws many samples from a population, given
population and sample characteristics supplied by the user, or else one of
twelve possible cases designed to cover all methods used in \code{ASL_table()}.

The sampling distributions are approximated by simulation for all estimators of age
proportions, abundances, and mean lengths, and their respective standard
errors. The sampling distributions are then graphically compared to true
values (in the case of simulated point estimates) or empirical standard deviations of
simulated point estimates (in the case of simulated standard errors.)

It should be noted that all simulated estimation is performed using the
current version of the \code{ASL_table()} function; therefore any bias or errors
that are identified should indicate a bug or incorrect derivation!

Twelve pre-determined cases may be used in the \verb{case=} argument, depending on
whether a stratified sampling scheme is used, whether abundance is known or
estimated with error, and whether estimates pertaining to categorical (age)
and/or numeric (length) values is to be performed.
}
\examples{
par(mfrow=c(2,2))
verify_ASL_table(case="stratified_witherror_lengthage", nsim=1000)


\dontrun{
nsim <- 5000
cases <- c("stratified_witherror_lengthage", "stratified_witherror_age",
"stratified_witherror_length", "stratified_lengthage", "stratified_age",
"stratified_length", "pooled_witherror_lengthage", "pooled_witherror_age",
"pooled_witherror_length" "pooled_lengthage", "pooled_age", "pooled_length")
for(case_i in cases) {
  par(mfrow=c(3,2))
  verify_ASL_table(case=case_i, nsim=5000)
}
}
}
\author{
Matt Tyers
}
