% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asl_func.R
\name{ASL_table}
\alias{ASL_table}
\title{Construct an ASL Table from Vectors of Data}
\usage{
ASL_table(
  age = NULL,
  sex = NULL,
  length = NULL,
  stratum = NULL,
  Nhat = NULL,
  se_Nhat = NULL,
  stratum_weights = NULL,
  verbose = FALSE,
  FPC = NA
)
}
\arguments{
\item{age}{Vector of ages.  If the default (\code{NULL}) is accepted, age will be
omitted from the resulting table.}

\item{sex}{Vector of sex.  If the default (\code{NULL}) is accepted, sex will be
omitted from the resulting table.}

\item{length}{Vector of length.  If the default (\code{NULL}) is accepted, length
will be omitted from the resulting table.}

\item{stratum}{Optional vector of stratum, if stratified estimators will be
used.  This must be formatted as positive whole numbers (1, 2, 3, etc).
Defaults to \code{NULL}.}

\item{Nhat}{Optional vector of abundance for each stratum, if stratified
estimators will be used.  The length of this vector must correspond to the
maximum value in the \verb{stratum=} argument.  Defaults to \code{NULL}.}

\item{se_Nhat}{Optional vector of the standard error of abundance for each
stratum, if stratified estimators will be used and by-stratum abundance is
considered to have measurement error.    The length of this vector must
correspond to the maximum value in the \verb{stratum=} argument.  Defaults to \code{NULL}.}

\item{stratum_weights}{Optional vector of weights for each stratum, if
relative weights are known but abundance is not.  Using this argument rather
than \code{Nhat} will also omit the Finite Population Correction factor (FPC) from
calculations, but weights are treated as constant (without error), likely
underestimating variance.  Defaults to \code{NULL}.}

\item{verbose}{Whether to print messages corresponding to the method used.
Defaults to \code{FALSE}.}

\item{FPC}{Whether to incorporate the Finite Population Correction factor in
variance calculations.  If the default (\code{NA}) is accepted, the FPC will only be
used if the abundance is considered to be known without error.}
}
\value{
A data.frame with rows corresponding to categories of age and/or sex,
depending on data inputs.
\subsection{Stratified}{
\subsection{Without error in Nhat}{
\subsection{If there are proportions}{

The proportion of each age and/or sex category will be estimated by sampling stratum (period) as follows:

\eqn{\hat{p}_{tz}=\frac{n_{tz}}{n_t}}

in which
\itemize{
\item \eqn{\hat{p}_{tz}} is the estimated proportion of fish during sampling stratum \emph{t} from age and/or sex category $z$
\item \eqn{n_{tz}} equals the number of fish sampled during sampling stratum $t$ that were classified as age and/or sex category $z$, and
\item \eqn{n_t} equals the number of fish sampled for sex determination during sampling stratum $t$.
}

Here is text
}

}

}
}
\description{
This function is intended to perform all the necessary calculations
to construct tables summarizing ASL (Age, Sex, and Length) for several data
scenarios:
\itemize{
\item Stratified vs. non-stratified
\item Abundance known without error vs. estimated with error vs. unknown
\item Data available for Age, Sex, and Length vs. some subset.
}

The function will return a table with rows for each combination of Age and
Sex (as available), and will summarize the respective proportions, estimated
abundance, and associated Lengths (as available).

Generally, \code{NULL} values in a given data vector indicate that the vector will
not be used.
}
\examples{
### a look at input
head(sim_data$data)
sim_data$abundance

### tables for possible scenarios

# Stratified with error in Nhat
ASL_table(age=sim_data$data$age,
          length=sim_data$data$length,
          stratum=sim_data$data$stratum,
          Nhat=sim_data$abundance$Nhat,
          se_Nhat=sim_data$abundance$se_Nhat)

# Stratified without error in Nhat
ASL_table(age=sim_data$data$age,
          length=sim_data$data$length,
          stratum=sim_data$data$stratum,
          Nhat=sim_data$abundance$Nhat)

# Pooled (not stratified) with error in Nhat
ASL_table(age=sim_data$data$age,
          length=sim_data$data$length,
          Nhat=100*1000,
          se_Nhat=10*1000)

# Pooled (not stratified) without error in Nhat
ASL_table(age=sim_data$data$age,
          length=sim_data$data$length,
          Nhat=100*1000)
}
\seealso{
\link{verify_ASL_table}
}
\author{
Matt Tyers
}
