% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asl_func.R
\name{ASL_table}
\alias{ASL_table}
\title{Construct an ASL Table from Vectors of Data}
\usage{
ASL_table(
  age = NULL,
  sex = NULL,
  length = NULL,
  stratum = NULL,
  Nhat = NULL,
  se_Nhat = NULL,
  stratum_weights = NULL,
  verbose = FALSE
)
}
\arguments{
\item{age}{Vector of ages.  If the default (\code{NULL}) is accepted, age will be
omitted from the resulting table.}

\item{sex}{Vector of sex.  If the default (\code{NULL}) is accepted, sex will be
omitted from the resulting table.}

\item{length}{Vector of length.  If the default (\code{NULL}) is accepted, length
will be omitted from the resulting table.}

\item{stratum}{Optional vector of stratum, if stratified estimators will be
used.  This must be formatted as positive whole numbers (1, 2, 3, etc).
Defaults to \code{NULL}.}

\item{Nhat}{Optional vector of abundance for each stratum, if stratified
estimators will be used.  The length of this vector must correspond to the
maximum value in the \verb{stratum=} argument.  Defaults to \code{NULL}.}

\item{se_Nhat}{Optional vector of the standard error of abundance for each
stratum, if stratified estimators will be used and by-stratum abundance is
considered to have measurement error.    The length of this vector must
correspond to the maximum value in the \verb{stratum=} argument.  Defaults to \code{NULL}.}

\item{stratum_weights}{Optional vector of weights for each stratum, if
relative weights are known but abundance is not.  Using this argument rather
than \code{Nhat} will also omit the Finite Population Correction factor (FPC) from
calculations, but weights are treated as constant (without error), likely
underestimating variance.  Defaults to \code{NULL}.}

\item{verbose}{Whether to print messages corresponding to the method used.
Defaults to \code{FALSE}.}
}
\value{
A data.frame with rows corresponding to categories of age and/or sex,
depending on data inputs.
}
\description{
FILL IN DESCRIPTION HERE
}
\author{
Matt Tyers
}
