% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asl_func.R
\name{ASL_table}
\alias{ASL_table}
\title{Construct an ASL Table from Vectors of Data}
\usage{
ASL_table(
  age = NULL,
  sex = NULL,
  length = NULL,
  stratum = NULL,
  Nhat = NULL,
  se_Nhat = NULL,
  stratum_weights = NULL,
  verbose = FALSE,
  FPC = c("ifknown", "always", "never")
)
}
\arguments{
\item{age}{Vector of ages.  If the default (\code{NULL}) is accepted, age will be
omitted from the resulting table.}

\item{sex}{Vector of sex.  If the default (\code{NULL}) is accepted, sex will be
omitted from the resulting table.}

\item{length}{Vector of length.  If the default (\code{NULL}) is accepted, length
will be omitted from the resulting table.}

\item{stratum}{Optional vector of stratum, if stratified estimators will be
used.  This must be formatted as positive whole numbers (1, 2, 3, etc).
Defaults to \code{NULL}.}

\item{Nhat}{Optional vector of abundance for each stratum, if stratified
estimators will be used.  The length of this vector must correspond to the
maximum value in the \verb{stratum=} argument.  Defaults to \code{NULL}.}

\item{se_Nhat}{Optional vector of the standard error of abundance for each
stratum, if stratified estimators will be used and by-stratum abundance is
considered to have measurement error.    The length of this vector must
correspond to the maximum value in the \verb{stratum=} argument.  Defaults to \code{NULL}.}

\item{stratum_weights}{Optional vector of weights for each stratum, if
relative weights are known but abundance is not.  Using this argument rather
than \code{Nhat} will also omit the Finite Population Correction factor (FPC) from
calculations, but weights are treated as constant (without error), likely
underestimating variance.  Defaults to \code{NULL}.}

\item{verbose}{Whether to print messages corresponding to the method used.
Defaults to \code{FALSE}.}

\item{FPC}{Whether to incorporate the Finite Population Correction factor (FPC) in
variance calculations.  Allowed values are
\itemize{
\item \code{"ifknown"} (the default), which will use the FPC only if abundance is known without error
\item \code{"always"}, which will use the FPC wherever possible (i.e. if there is an estimate of abundance)
\item \code{"never"}, which will always ignore the FPC.
used if the abundance is considered to be known without error.
}}
}
\value{
A data.frame with rows corresponding to categories of age and/or sex,
depending on data inputs.
\subsection{Stratified - If abundance is known without error}{
\subsection{If there are proportions}{

The proportion of each age and/or sex category \emph{z} will be estimated for each sampling stratum \emph{t} as follows:

\deqn{\hat{p}_{tz}=\frac{n_{tz}}{n_t}}

in which \eqn{n_{tz}} equals the number of fish sampled during sampling stratum \eqn{t} classified as age and/or sex category \eqn{z}, and \eqn{n_t} equals the number of fish sampled for age and/or sex determination within sampling stratum \eqn{t}.

The sampling variance of \eqn{\hat{p}_{tz}} will be estimated as the following (Cochran 1977), in which \eqn{N_t} represents the total abundance of fish in sampling stratum \emph{t}:

\deqn{\hat{var}[\hat{p}_{tz}]=\frac{\hat{p}_{tz}(1-\hat{p}_{tz})}{n_t-1}\left(\frac{N_t-n_t}{N_t-1}\right)}

if the finite population correction factor (FPC) is used; otherwise, as the following:

\deqn{\hat{var}[\hat{p}_{tz}]=\frac{\hat{p}_{tz}(1-\hat{p}_{tz})}{n_t-1}}

The total abundance by age and/or sex category in each sampling stratum will be estimated as follows:

\deqn{\hat{N}_{tz}=N_t\hat{p}_{tz}}

with variance estimated as

\deqn{\hat{var}[\hat{N}_{tz}]=N_t^2\hat{var}[\hat{p}_{tz}]}

The total abundance by age and/or sex category and its variance will then be estimated by summation as follows:

\deqn{\hat{N}_z=\sum_{t=1}^{L}\hat{N}_{tz}}

and

\deqn{\hat{var}[\hat{N}_{z}]=\sum_{t=1}^{L}\hat{var}[\hat{N}_{tz}]}

where \eqn{L} equals the number of sampling strata.

Finally, the overall proportion by age and/or sex category and its variance will be estimated as follows:

\deqn{\hat{p}_z=\frac{\hat{N}_z}{N}}

and

\deqn{\hat{var}[\hat{p}_z]=\frac{\hat{var}[\hat{N}_z]}{N^2}}

where \eqn{N} is the total abundance across all sampling periods.

The mean length by age and/or sex for each sampling stratum will be estimated as follows:

\deqn{\bar{x}_{tz}=\frac{\sum_{i=1}^{n_{tz}}x_{tzi}}{n_{tz}}}

where \eqn{x_{tzi}} is the length of the \emph{i}th fish sampled of age and/or sex \eqn{z} during sampling stratum \eqn{t}.

The sampling variance of \eqn{\bar{x}_{tz}} will be estimated as

\deqn{\hat{var}[\bar{x}_{tz}]=\frac{\sum_{i=1}^{n_{tz}}(x_{tzi}-\bar{x}_{tz})^2}{n_{tz}(n_{tz}-1)}\left(\frac{
    \hat{N}_{tz}-n_{tz}}{\hat{N}_{tz}-1}\right)}

if the finite population correction factor (FPC) will be used; otherwise, as the following:

\deqn{\hat{var}[\bar{x}_{tz}]=\frac{\sum_{i=1}^{n_{tz}}(x_{tzi}-\bar{x}_{tz})^2}{n_{tz}(n_{tz}-1)}}

The mean length by age and/or sex category will then be estimated as follows:

\deqn{\bar{x}_z=\sum_{t=1}^{L}\frac{\hat{N}_{tz}}{\hat{N}_z}\bar{x}_{tz}}

with its variance approximated using a Taylor's series expansion (Mood et al. 1974):

\deqn{\hat{var}[\bar{x}_z]\approx\sum_{t=1}^{L}\frac{\hat{N}_{tz}^2}{\hat{N}_z^2}\hat{var}[\bar{x}_{tz}]+\sum_{t=1}^{L}\frac{\left(\bar{x}_{tz}\hat{N}_z-\left(\sum_{u=1}^{L}\bar{x}_{uz}\hat{N}_{uz}\right)\right)^2}{\hat{N}_z^4}\hat{var}[\hat{N}_{tz}]}
}

\subsection{If there are no proportions to estimate}{

The mean length for each sampling stratum will be estimated as follows, where \eqn{x_{ti}} is the length of the \emph{i}th fish sampled within sampling stratum \eqn{t}, and \eqn{n_t} is the number of fish in stratum \emph{t} sampled for length:

\deqn{\bar{x}_{t}=\frac{\sum_{i=1}^{n_{t}}x_{ti}}{n_{t}}}

The sampling variance of \eqn{\bar{x}_{t}} will be estimated as

\deqn{\hat{var}[\bar{x}_{t}]=\frac{\sum_{i=1}^{n_{t}}(x_{ti}-\bar{x}_{t})^2}{n_{t}(n_{t}-1)}\left(\frac{
N_{t}-n_{t}}{N_{t}-1}\right)}

if abundance per stratum \eqn{N_t} is known and if the finite population correction factor is used, otherwise as:

\deqn{\hat{var}[\bar{x}_{t}]=\frac{\sum_{i=1}^{n_{t}}(x_{ti}-\bar{x}_{t})^2}{n_{t}(n_{t}-1)}}

Stratified estimates of mean length will be calculated as follows, in which \eqn{N_t} represents the abundance associated with sampling stratum \eqn{t}, \eqn{N} represents the total abundance, and \emph{L} represents the number of sampling strata:

\deqn{\bar{x}=\frac{1}{N}\sum_{t=1}^L N_t\bar{x}_t}

and

\deqn{\hat{var}[\bar{x}]=\sum_{t=1}^L\left(\frac{N_t}{N}\right)^2 \hat{var}[\bar{x}_t]}
}

}

\subsection{Stratified - If abundance is estimated with error}{
\subsection{If there are proportions}{

The proportion of each age and/or sex category \emph{z} will be estimated for each sampling stratum \emph{t} as follows:

\deqn{\hat{p}_{tz}=\frac{n_{tz}}{n_t}}

in which \eqn{n_{tz}} equals the number of fish sampled during sampling stratum \eqn{t} classified as age and/or sex category \eqn{z}, and \eqn{n_t} equals the number of fish sampled for age and/or sex determination within sampling stratum \eqn{t}.

The sampling variance of \eqn{\hat{p}_{tz}} will be estimated as the following (Cochran 1977) in which \eqn{\hat{N}_t} is the estimated abundance of fish in sampling stratum \eqn{t}:

\deqn{\hat{var}[\hat{p}_{tz}]=\frac{\hat{p}_{tz}(1-\hat{p}_{tz})}{n_t-1}\left(\frac{\hat{N}_t-n_t}{\hat{N}_t-1}\right)}

if the finite population correction factor (FPC) is used; otherwise, as the following:

\deqn{\hat{var}[\hat{p}_{tz}]=\frac{\hat{p}_{tz}(1-\hat{p}_{tz})}{n_t-1}}

The total abundance by age and/or sex category in each sampling stratum will be estimated as follows:

\deqn{\hat{N}_{tz}=\hat{N}_t\hat{p}_{tz}}

with variance estimated as (Goodman 1960):

\deqn{\hat{var}[\hat{N}_{tz}]=\hat{N}_t^2\hat{var}[\hat{p}_{tz}] + \hat{p}_{tz}^2\hat{var}[\hat{N}_t]-\hat{var}[\hat{p}_{tz}]\hat{var}[\hat{p}_{tz}]}

The total abundance by age and/or sex category \eqn{z} and its variance will then be estimated by summation as follows:

\deqn{\hat{N}_z=\sum_{t=1}^{L}\hat{N}_{tz}}

and

\deqn{\hat{var}[\hat{N}_{z}]=\sum_{t=1}^{L}\hat{var}[\hat{N}_{tz}]}

where \eqn{L} equals the number of sampling strata.

Finally, the overall proportion by age and/or sex category and its variance will be estimated as follows:

\deqn{\hat{p}_z=\frac{\hat{N}_z}{\sum_{t=1}^{L}\hat{N}_t}}

with variance estimated by the delta method (Casella & Berger 2002) as:

\deqn{\hat{var}[\hat{p}_z] \approx \left(\frac{\hat{N}_z}{\sum_{t=1}^{L}\hat{N}_t}\right)^2\left(\frac{\hat{var}[\hat{N}_z]}{\hat{N}_z^2} + \frac{\sum_{t=1}^{L}\hat{var}[\hat{N}_t]}{(\sum_{t=1}^{L}\hat{N}_t)^2} - 2\frac{\hat{cov}[\hat{N}_z,\sum_{t=1}^{L}\hat{N}_t]}{\hat{N}_z\sum_{t=1}^{L}\hat{N}_t}\right)}

in which

\deqn{\hat{cov}[\hat{N}_z,\sum_{t=1}^{L}\hat{N}_t]=\sum_{t=1}^{L}\hat{p}_{tz}\hat{var}[\hat{N_t}]}

The mean length by age and/or sex for each sampling stratum will be estimated as follows:

\deqn{\bar{x}_{tz}=\frac{\sum_{i=1}^{n_{tz}}x_{tzi}}{n_{tz}}}

where \eqn{x_{tzi}} is the length of the \emph{i}th fish sampled of age and/or sex \eqn{z} during sampling stratum \eqn{t}.

The sampling variance of \eqn{\bar{x}_{tz}} will be estimated as

\deqn{\hat{var}[\bar{x}_{tz}]=\frac{\sum_{i=1}^{n_{tz}}(x_{tzi}-\bar{x}_{tz})^2}{n_{tz}(n_{tz}-1)}\left(\frac{\hat{N}_{tz}-n_{tz}}{\hat{N}_{tz}-1}\right)}

if the finite population correction factor (FPC) is used; otherwise, as

\deqn{\hat{var}[\bar{x}_{tz}]=\frac{\sum_{i=1}^{n_{tz}}(x_{tzi}-\bar{x}_{tz})^2}{n_{tz}(n_{tz}-1)}}

The mean length by age and/or sex category will then be estimated as follows:

\deqn{\bar{x}_z=\sum_{t=1}^{L}\frac{\hat{N}_{tz}}{\hat{N}_z}\bar{x}_{tz}}

with its variance approximated using a Taylor's series expansion (Mood et al. 1974):

\deqn{\hat{var}[\bar{x}_z]\approx\sum_{t=1}^{L}\frac{\hat{N}_{tz}^2}{\hat{N}_z^2}\hat{var}[\bar{x}_{tz}]+\sum_{t=1}^{L}\frac{\left(\bar{x}_{tz}\hat{N}_z-\left(\sum_{u=1}^{L}\bar{x}_{uz}\hat{N}_{uz}\right)\right)^2}{\hat{N}_z^4}\hat{var}[\hat{N}_{tz}]}
}

\subsection{If there are no proportions to estimate}{

The mean length for each sampling stratum will be estimated as follows, where \eqn{x_{ti}} is the length of the \emph{i}th fish sampled within sampling stratum \eqn{t}, and \eqn{n_t} is the number of fish in stratum \emph{t} sampled for length:

\deqn{\bar{x}_{t}=\frac{\sum_{i=1}^{n_{t}}x_{ti}}{n_{t}}}

The sampling variance of \eqn{\bar{x}_{t}} will be estimated as

\deqn{\hat{var}[\bar{x}_{t}]=\frac{\sum_{i=1}^{n_{t}}(x_{ti}-\bar{x}_{t})^2}{n_{t}(n_{t}-1)}\left(\frac{
\hat{N}_{t}-n_{t}}{\hat{N}_{t}-1}\right)}

if the finite population correction factor (FPC) is used, otherwise as:

\deqn{\hat{var}[\bar{x}_{t}]=\frac{\sum_{i=1}^{n_{t}}(x_{ti}-\bar{x}_{t})^2}{n_{t}(n_{t}-1)}}

Stratified estimates of mean length will be calculated as follows, in which \eqn{\hat{N}_t} and \eqn{\bar{x}_t} represent the estimated abundance and mean length associated with stratum \emph{t}, respectively:

\deqn{\bar{x}=\frac{\sum_{t=1}^L \hat{N}_t\bar{x}_t}{\sum_{t=1}^L \hat{N}_t}}

and

\deqn{\hat{var}[\bar{x}] \approx \left(\frac{\sum_{t=1}^L \hat{N}_t\bar{x}_t}{\sum_{t=1}^L \hat{N}_t}\right)^2\left(\frac{\hat{var}[\sum_{t=1}^L \hat{N}_t\bar{x}_t]}{\left(\sum_{t=1}^L \hat{N}_t\bar{x}_t\right)^2}+\frac{\sum_{t=1}^L \hat{var}[\hat{N}_t]}{\left(\sum_{t=1}^L \hat{N}_t\right)^2}-2\frac{\hat{cov}[\sum_{t=1}^L \hat{N}_t,\sum_{t=1}^L N_t\bar{x}_t]}{\left(\sum_{t=1}^L \hat{N}_t\right)\left(\sum_{t=1}^L \hat{N}_t\bar{x}_t\right)}\right)}

in which

\deqn{\hat{cov}[\sum_{t=1}^L \hat{N}_t,\sum_{t=1}^L \hat{N}_t\bar{x}_t]=\sum_{t=1}^L \bar{x}_t\hat{var}[\hat{N}_t]}

and

\deqn{\hat{var}[\sum_{t=1}^L \hat{N}_t\bar{x}_t]=\sum_{t=1}^L\hat{N}_t^2\hat{var}[\bar{x}_t] + \bar{x}_t^2\hat{var}[\hat{N}_t]-\hat{var}[\hat{N}_t]\hat{var}[\bar{x}_t]}

by means of the delta method (Casella & Berger 2002) and Goodman (1960), respectively.
}

}

\subsection{Stratified - If abundance is unknown and sample weights are used}{
\subsection{If there are proportions}{

The proportion of each age and/or sex category \emph{z} will be estimated for each sampling stratum \emph{t} as follows:

\deqn{\hat{p}_{tz}=\frac{n_{tz}}{n_t}}

in which \eqn{n_{tz}} equals the number of fish sampled during sampling stratum \eqn{t} classified as age and/or sex category \eqn{z}, and \eqn{n_t} equals the number of fish sampled for age and/or sex determination within sampling stratum \eqn{t}.

The sampling variance of \eqn{\hat{p}_{tz}} will be estimated as the following (Cochran 1977):

\deqn{\hat{var}[\hat{p}_{tz}]=\frac{\hat{p}_{tz}(1-\hat{p}_{tz})}{n_t-1}}

The overall proportion by age and/or sex category and its variance will be estimated as follows, in which \eqn{w_t} represents the sampling weight associated with stratum \emph{t} and \emph{L} equals the number of strata.  It is worth noting that weights \eqn{w_t} are treated as constant (i.e. known without error), therefore all variance estimates must be interpreted as minima without further assumptions.

\deqn{\hat{p}_z=\frac{\sum_{t=1}^Lw_t\hat{p}_{tz}}{\sum_{t=1}^Lw_t}}

and

\deqn{\hat{var}[\hat{p}_z]=\frac{\sum_{t=1}^Lw_t^2\hat{var}[\hat{p}_{tz}]}{\left(\sum_{t=1}^Lw_t\right)^2}}

The mean length by age and/or sex for each sampling stratum will be estimated as follows:

\deqn{\bar{x}_{tz}=\frac{\sum_{i=1}^{n_{tz}}x_{tzi}}{n_{tz}}}

where \eqn{x_{tzi}} is the length of the \emph{i}th fish sampled of age and/or sex \eqn{z} during sampling stratum \eqn{t}.

The sampling variance of \eqn{\bar{x}_{tz}} will be estimated as:

\deqn{\hat{var}[\bar{x}_{tz}]=\frac{\sum_{i=1}^{n_{tz}}(x_{tzi}-\bar{x}_{tz})^2}{n_{tz}(n_{tz}-1)}}

The mean length by age and/or sex category will then be estimated as follows:

\deqn{\bar{x}_z=\frac{\sum_{t=1}^{L}w_t\hat{p}_{tz}\bar{x}_{tz}}{\sum_{t=1}^{L}w_t\hat{p}_{tz}}}

with its variance approximated using a Taylor's series expansion (Mood et al. 1974):

\deqn{\hat{var}[\bar{x}_z]\approx\frac{\sum_{t=1}^{L}w_t\hat{p}_{tz}^2\hat{var}[\bar{x}_{tz}]}{\left(\sum_{t=1}^Lw_t\hat{p}_{tz}\right)^2}+\frac{\sum_{t=1}^{L}\left(\bar{x}_{tz}\sum_{u=1}^Lw_u\hat{p}_{uz}-\left(\sum_{u=1}^{L}\bar{x}_{uz}w_u\hat{p}_{uz}\right)\right)^2w_t^2\hat{var}[\hat{p}_{tz}]}{\left(\sum_{t=1}^Lw_t\hat{p}_{tz}\right)^4}}
}

\subsection{If there are no proportions to estimate}{

The mean length for each sampling stratum will be estimated as follows, where \eqn{x_{ti}} is the length of the \emph{i}th fish sampled within sampling stratum \eqn{t}, and \eqn{n_t} is the number of fish in stratum \emph{t} sampled for length:

\deqn{\bar{x}_{t}=\frac{\sum_{i=1}^{n_{t}}x_{ti}}{n_{t}}}

The sampling variance of \eqn{\bar{x}_{t}} will be estimated as:

\deqn{\hat{var}[\bar{x}_{t}]=\frac{\sum_{i=1}^{n_{t}}(x_{ti}-\bar{x}_{t})^2}{n_{t}(n_{t}-1)}}

Stratified estimates of mean length will be calculated as follows, in which \eqn{w_t} and \eqn{\bar{x}_t} represent the sampling weight and average length associated with sampling stratum \eqn{t}, respectively.  It is worth noting that weights \eqn{w_t} are treated as constant (i.e. known without error), therefore all variance estimates must be interpreted as minima without further assumptions.

\deqn{\bar{x}=\frac{\sum_{t=1}^L w_t\bar{x}_t}{\sum_{t=1}^L w_t}}

and

\deqn{\hat{var}[\bar{x}]=\frac{\sum_{t=1}^L w_t^2\hat{var}[\bar{x}_t]}{\left(\sum_{t=1}^L w_t\right)^2}}
}

}

\subsection{Pooled (not stratified) - If abundance is known without error}{
\subsection{If there are proportions}{

#' Proportions of each age and/or sex category \eqn{z} will be estimated as follows (Cochran 1977):

\deqn{\hat{p}_z=\frac{n_z}{n}}

and

\deqn{\hat{var}[\hat{p}_z]=\frac{\hat{p}(1-\hat{p})}{n-1}\left(\frac{N-n}{N-1}\right)}

if total abundance \eqn{N} is known and the finite population correction factor (FPC) is used; otherwise as

\deqn{\hat{var}[\hat{p}_z]=\frac{\hat{p}_z(1-\hat{p}_z)}{n-1}}

in which \eqn{n_z} denotes the number of fish sampled in age and/or sex category \eqn{z}, \eqn{n} denotes the total number of fish sampled, and \emph{N} denotes the total abundance.

Total abundance for age and/or sex category \eqn{z} will be estimated as

\deqn{\hat{N}_z=N\hat{p}_z} and

\deqn{\hat{var}[\hat{N}_z]=N^2\hat{var}[\hat{p}_z]}

The mean length associated with age and/or sex category \eqn{z} will be estimated as the following, in which \eqn{x_{zi}} represents the length of the \emph{i}th fish in age and/or sex category \emph{z} and \eqn{n_z} represents the number of fish in age and/or sex category \emph{z} with an associated length measurement:

\deqn{\bar{x}_z=\frac{\sum_{i=1}^{n_z}x_{zi}}{n_z}}

and

\deqn{\hat{var}[\bar{x}_z]=\frac{\sum_{i=1}^{n_z}(x_{zi}-\bar{x_z})^2}{n_z(n_z-1)}\left(\frac{\hat{N}_z-n_z}{\hat{N}_z-1}\right)}

if the finite population correction factor (FPC) is used, otherwise as:

\deqn{\hat{var}[\bar{x}_z]=\frac{\sum_{i=1}^{n_z}(x_{zi}-\bar{x_z})^2}{n_z(n_z-1)}}
}

\subsection{If there are no proportions to estimate}{

The mean length of all fish will be estimated as the following, in which \eqn{x_{i}} represents the length of the \emph{i}th fish, \emph{n} represents the number of fish with an associated length measurement, and \emph{N} represents the total abundance:

\deqn{\bar{x}=\frac{\sum_{i=1}^{n}x_{i}}{n}}

and

\deqn{\hat{var}[\bar{x}]=\frac{\sum_{i=1}^{n}(x_{i}-\bar{x})^2}{n(n-1)}\left(\frac{N-n}{N-1}\right)}

if the finite population correction factor (FPC) is used, otherwise as:

\deqn{\hat{var}[\bar{x}]=\frac{\sum_{i=1}^{n}(x_{i}-\bar{x})^2}{n(n-1)}}
}

}

\subsection{Pooled (not stratified) - If abundance is estimated with error}{
\subsection{If there are proportions}{

Proportions of each age and/or sex category \eqn{z} will be estimated as follows (Cochran 1977):

\deqn{\hat{p}_z=\frac{n_z}{n}}

and

\deqn{\hat{var}[\hat{p}]=\frac{\hat{p}(1-\hat{p})}{n-1}\left(\frac{\hat{N}-n}{\hat{N}-1}\right)}

if the finite population correction factor (FPC) is used, otherwise as

\deqn{\hat{var}[\hat{p}]=\frac{\hat{p}(1-\hat{p})}{n-1}}

in which \eqn{n_z} denotes the number of fish sampled in age and/or sex category \eqn{z}, \eqn{n} denotes the total number of fish sampled, and \eqn{\hat{N}} denotes the estimated abundance.

Total abundance for age and/or sex category \eqn{z} will be estimated as follows (Goodman 1960):

\deqn{\hat{N}_z=\hat{N}\hat{p}_z}

and

\deqn{\hat{var}[\hat{N}_z]=\hat{N}^2\hat{var}[\hat{p}_z] + \hat{p}_z^2\hat{var}[\hat{N}]-\hat{var}[\hat{N}]\hat{var}[\hat{p}_z]}

The mean length associated with age and/or sex category \eqn{z} will be estimated as the following, in which \eqn{x_{zi}} represents the length of fish \emph{i} within age and/or sex category category \emph{z}, and \eqn{n_z} denotes the number of fish within age and/or sex category \emph{z} with an associated length measurement:

\deqn{\bar{x}_z=\frac{\sum_{i=1}^{n_z}x_{zi}}{n_z}}

and

\deqn{\hat{var}[\bar{x}_z]=\frac{\sum_{i=1}^{n_z}(x_{zi}-\bar{x_z})^2}{n_z(n_z-1)}\left(\frac{\hat{N}_z-n_z}{\hat{N}_z-1}\right)}

if the finite population correction factor (FPC) is used; otherwise as

\deqn{\hat{var}[\bar{x}_z]=\frac{\sum_{i=1}^{n_z}(x_{zi}-\bar{x_z})^2}{n_z(n_z-1)}}
}

\subsection{If there are no proportions to estimate}{

The mean length of all fish will be estimated as the following, in which \eqn{x_{i}} represents the length of the \emph{i}th fish, \emph{n} represents the number of fish with an associated length measurement, and \eqn{\hat{N}} represents the estimated total abundance:

\deqn{\bar{x}=\frac{\sum_{i=1}^{n}x_{i}}{n}}

and

\deqn{\hat{var}[\bar{x}]=\frac{\sum_{i=1}^{n}(x_{i}-\bar{x})^2}{n(n-1)}\left(\frac{\hat{N}-n}{\hat{N}-1}\right)}

if the finite population correction factor (FPC) is used, otherwise as:

\deqn{\hat{var}[\bar{x}]=\frac{\sum_{i=1}^{n}(x_{i}-\bar{x})^2}{n(n-1)}}
}

}

\subsection{Pooled (not stratified) - If abundance is unknown}{
\subsection{If there are proportions}{

Proportions of each age and/or sex category \eqn{z} will be estimated as follows (Cochran 1977):

\deqn{\hat{p}_z=\frac{n_z}{n}}

and

\deqn{\hat{var}[\hat{p}_z]=\frac{\hat{p}_z(1-\hat{p}_z)}{n-1}}

in which \eqn{n_z} denotes the number of fish sampled in age and/or sex category \eqn{z}, and \eqn{n} denotes the total number of fish sampled.

The mean length associated with age and/or sex category \eqn{z} will be estimated as the following, in which \eqn{x_{zi}} denotes the length measurement associated with the \emph{i}th fish in age and/or sex category \emph{z}:

\deqn{\bar{x}_z=\frac{\sum_{i=1}^{n_z}x_{zi}}{n_z}}

and

\deqn{\hat{var}[\bar{x}_z]=\frac{\sum_{i=1}^{n_z}(x_{zi}-\bar{x_z})^2}{n_z(n_z-1)}}
}

\subsection{If there are no proportions to estimate}{

The mean length will be estimated as the following, in which \eqn{x_{i}} denotes the length measurement associated with the \emph{i}th fish and \emph{n} denotes the number of fish with an associated length measurement:

\deqn{\bar{x}=\frac{\sum_{i=1}^{n}x_{i}}{n}}

and

\deqn{\hat{var}[\bar{x}]=\frac{\sum_{i=1}^{n}(x_{i}-\bar{x})^2}{n(n-1)}}
}

}
}
\description{
This function is intended to perform all the necessary calculations
to construct tables summarizing ASL (Age, Sex, and Length) for several data
scenarios:
\itemize{
\item Stratified vs. non-stratified
\item Abundance known without error vs. estimated with error vs. unknown
\item Data available for Age, Sex, and Length vs. some subset.
}

The function will return a table with rows for each combination of Age and
Sex (as available), and will summarize the respective proportions, estimated
abundance, and associated Lengths (as available).

Generally, \code{NULL} values in a given data vector indicate that the vector will
not be used.

Methods are also described in the document found here:

\verb{https://github.com/ADFG-DSF/dsftools/blob/main/addl_documentation/asl_equations.Rmd}
}
\examples{
### a look at input
head(sim_data$data)
sim_data$abundance

### tables for possible scenarios

# Stratified with error in Nhat
ASL_table(age=sim_data$data$age,
          length=sim_data$data$length,
          stratum=sim_data$data$stratum,
          Nhat=sim_data$abundance$Nhat,
          se_Nhat=sim_data$abundance$se_Nhat)

# Stratified without error in Nhat
ASL_table(age=sim_data$data$age,
          length=sim_data$data$length,
          stratum=sim_data$data$stratum,
          Nhat=sim_data$abundance$Nhat)

# Pooled (not stratified) with error in Nhat
ASL_table(age=sim_data$data$age,
          length=sim_data$data$length,
          Nhat=100*1000,
          se_Nhat=10*1000)

# Pooled (not stratified) without error in Nhat
ASL_table(age=sim_data$data$age,
          length=sim_data$data$length,
          Nhat=100*1000)
}
\references{
Casella, George and Roger L. Berger. 2002. \emph{Statistical Inference}. Australia ; Pacific Grove, CA, Thomson Learning

Cochran, W. G.  1977.  \emph{Sampling techniques}. 3rd edition.  John Wiley and Sons, New York.

Goodman, L.A., 1960. On the exact variance of products. \emph{Journal of the American statistical association, 55}(292), pp.708-713.

Mood, A. M., F. A. Graybill, and D. C. Boes.  1974.  \emph{Introduction to the theory of statistics}. 3rd edition.  McGraw-Hill Book Co., New York.
}
\seealso{
\link{verify_ASL_table}
}
\author{
Matt Tyers
}
