% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asl_sim.R
\name{rp_ASL_table}
\alias{rp_ASL_table}
\title{Relative precision for the Methods Used in \code{ASL_table()}}
\usage{
rp_ASL_table(
  conf_target = 0.9,
  sim = NULL,
  case = NULL,
  nstrata,
  nage,
  nt,
  Nt,
  stratum_weights = NULL,
  se_Nt = NULL,
  mn_length = NULL,
  sd_length = NULL,
  ptz = NULL,
  nNA = 0,
  nsim = 1000,
  plot_pop = TRUE,
  verbose = TRUE,
  print_table = FALSE
)
}
\arguments{
\item{conf_target}{The target level of confidence desired.  Defaults to \code{0.9}.}

\item{sim}{An object created by \link{simulate_ASL_table}.  If the default (\code{NULL})
is accepted, the simulation function will be called internally from the other
arguments provided.}

\item{case}{If a pre-determined case is to be used.  Allowed values are
\code{"stratified_witherror_lengthage"}, \code{"stratified_witherror_age"}, \code{"stratified_witherror_length"},
\code{"stratified_lengthage"}, \code{"stratified_age"}, \code{"stratified_length"},
\code{"stratified_Nunknown_lengthage"}, \code{"stratified_Nunknown_age"}, \code{"stratified_Nunknown_length"},
\code{"pooled_witherror_lengthage"}, \code{"pooled_witherror_age"}, \code{"pooled_witherror_length"}
\code{"pooled_lengthage"}, \code{"pooled_age"},  \code{"pooled_length"},
\code{"pooled_Nunknown_lengthage"}, \code{"pooled_Nunknown_length"}, or \code{"pooled_Nunknown_age"}. If the default
(\code{NULL}) is accepted, all simulation parameters below must be supplied by the user.}

\item{nstrata}{Number of sampling strata}

\item{nage}{Number of age categories}

\item{nt}{Sample size for each stratum}

\item{Nt}{Abundance for each stratum.  If abundance is completely unknown, sampling
weights may be used instead in the \verb{sampling_weights=} argument.}

\item{stratum_weights}{Optional vector of sampling weights for each stratum.  Defaults to \code{NULL}.}

\item{se_Nt}{Optional vector of standard errors of abundance for each stratum
Defaults to \code{NULL}, implying abundance is known without error.}

\item{mn_length}{Vector of mean lengths for each age category.  Defaults to
\code{NULL}, implying lengths were not considered.}

\item{sd_length}{Vector of standard deviations for each age category.  Defaults to
\code{NULL}, implying lengths were not considered.}

\item{ptz}{Matrix of probabilities of each age by stratum, with rows
corresponding to strata and columns corresponding to ages.  The probabilities
for each row will be normalized before simulation, therefore summing to one
is not required.  If a pooled (non-stratified) sample is to be taken and
age categories are to be considered, this should be supplied as a vector with
length equal to the number of ages.}

\item{nNA}{Number of NA values to randomly impute, to test robustness to NA.
Defaults to \code{0} in this case.}

\item{nsim}{Number of simulated replicates.  Defaults to \code{1000}, but more is recommended.}

\item{plot_pop}{Whether to make summary plots of the simulated population and
one representative sample, in addition to the plots produced in simulation.
Defaults to \code{TRUE}.}

\item{verbose}{Whether to print the parameters used in simulation to the console,
if one of the \code{case}s is accepted, as well as printing the method used within
\code{ASL_table()}.  Defaults to \code{TRUE}.}

\item{print_table}{Whether to print an example output table from \code{ASL_table()}
as an additional check.  Defaults to \code{FALSE}.}
}
\value{
\code{NULL}
}
\description{
This function may be used with a simulation object created by
\link{simulate_ASL_table}, or will call it internally if none is supplied.

The simulation function draws many samples from a population, given
population and sample characteristics supplied by the user, or else one of
twelve possible cases designed to cover all methods used in \code{ASL_table()}.

This function uses the simulation results to produce a sequence of plots
displaying the relationships between relative precision (how close we want
estimates to be to the true value(s)) and confidence (how often we can expect
estimates to be within the desired relative precision of the truth).

In terms of a typical Operational Plan Objective: "Estimate ... such that all
estimates are within (precision) of the true values (confidence) \% of the time".

Relative precision curves will be created for proportions (phat), abundance (Nhat),
and mean length (mn_length), depending on which are present.  Multiple curves
will be overlayed for each, if multiple categories are present for Age and/or Sex.
}
\examples{
par(mfrow=c(2,2))
verify_ASL_table(case="stratified_witherror_lengthage", nsim=1000)


\dontrun{
nsim <- 5000
cases <- c("stratified_witherror_lengthage", "stratified_witherror_age",
"stratified_witherror_length", "stratified_lengthage", "stratified_age",
"stratified_length", "stratified_Nunknown_lengthage", "stratified_Nunknown_age",
"stratified_Nunknown_length", "pooled_witherror_lengthage", "pooled_witherror_age",
"pooled_witherror_length", "pooled_lengthage", "pooled_age", "pooled_length",
"pooled_Nunknown_lengthage", "pooled_Nunknown_age", "pooled_Nunknown_length")
for(case_i in cases) {
  par(mfrow=c(3,2))
  verify_ASL_table(case=case_i, nsim=nsim)
}
}
}
\seealso{
\link{ASL_table}
}
\author{
Matt Tyers
}
