% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asl_sim.R
\name{verify_ASL_table}
\alias{verify_ASL_table}
\title{Empirical Verification of the Methods Used in \code{ASL_table()}}
\usage{
verify_ASL_table(
  sim = NULL,
  case = NULL,
  nstrata,
  nage,
  nt,
  Nt,
  stratum_weights = NULL,
  se_Nt = NULL,
  mn_length = NULL,
  sd_length = NULL,
  ptz = NULL,
  nNA = 10,
  nsim = 1000,
  plot_pop = TRUE,
  verbose = TRUE,
  print_table = FALSE
)
}
\arguments{
\item{sim}{An object created by \link{simulate_ASL_table}.  If the default (\code{NULL})
is accepted, the simulation function will be called internally from the other
arguments provided.}

\item{case}{If a pre-determined case is to be used.  Allowed values are
\code{"stratified_witherror_lengthage"}, \code{"stratified_witherror_age"}, \code{"stratified_witherror_length"},
\code{"stratified_lengthage"}, \code{"stratified_age"}, \code{"stratified_length"},
\code{"stratified_Nunknown_lengthage"}, \code{"stratified_Nunknown_age"}, \code{"stratified_Nunknown_length"},
\code{"pooled_witherror_lengthage"}, \code{"pooled_witherror_age"}, \code{"pooled_witherror_length"}
\code{"pooled_lengthage"}, \code{"pooled_age"},  \code{"pooled_length"},
\code{"pooled_Nunknown_lengthage"}, \code{"pooled_Nunknown_length"}, or \code{"pooled_Nunknown_age"}. If the default
(\code{NULL}) is accepted, all simulation parameters below must be supplied by the user.}

\item{nstrata}{Number of sampling strata}

\item{nage}{Number of age categories}

\item{nt}{Sample size for each stratum}

\item{Nt}{Abundance for each stratum.  If abundance is completely unknown, sampling
weights may be used instead in the \verb{sampling_weights=} argument.}

\item{stratum_weights}{Optional vector of sampling weights for each stratum.  Defaults to \code{NULL}.}

\item{se_Nt}{Optional vector of standard errors of abundance for each stratum
Defaults to \code{NULL}, implying abundance is known without error.}

\item{mn_length}{Vector of mean lengths for each age category.  Defaults to
\code{NULL}, implying lengths were not considered.}

\item{sd_length}{Vector of standard deviations for each age category.  Defaults to
\code{NULL}, implying lengths were not considered.}

\item{ptz}{Matrix of probabilities of each age by stratum, with rows
corresponding to strata and columns corresponding to ages.  The probabilities
for each row will be normalized before simulation, therefore summing to one
is not required.  If a pooled (non-stratified) sample is to be taken and
age categories are to be considered, this should be supplied as a vector with
length equal to the number of ages.}

\item{nNA}{Number of NA values to randomly impute, to test robustness to NA.  Defaults to \code{10}.}

\item{nsim}{Number of simulated replicates.  Defaults to \code{1000}, but more is recommended.}

\item{plot_pop}{Whether to make summary plots of the simulated population and
one representative sample, in addition to the plots produced in simulation.
Defaults to \code{TRUE}.}

\item{verbose}{Whether to print the parameters used in simulation to the console,
if one of the \code{case}s is accepted, as well as printing the method used within
\code{ASL_table()}.  Defaults to \code{TRUE}.}

\item{print_table}{Whether to print an example output table from \code{ASL_table()}
as an additional check.  Defaults to \code{FALSE}.}
}
\value{
\code{NULL}
}
\description{
This function may be used with a simulation object created by
\link{simulate_ASL_table}, or will call it internally if none is supplied.

The simulation function draws many samples from a population, given
population and sample characteristics supplied by the user, or else one of
twelve possible cases designed to cover all methods used in \code{ASL_table()}.

The sampling distributions are approximated by simulation for all estimators of age
proportions, abundances, and mean lengths, and their respective standard
errors. The sampling distributions are then graphically compared to true
values (in the case of simulated point estimates) or empirical standard deviations of
simulated point estimates (in the case of simulated standard errors.)

It should be noted that all simulated estimation is performed using the
current version of the \code{ASL_table()} function; therefore any bias or errors
that are identified should indicate a bug or incorrect derivation!

Eighteen pre-determined cases may be used in the \verb{case=} argument, depending on
whether a stratified sampling scheme is used, whether abundance is known,
estimated with error, or completely unknown, and whether estimates pertaining to categorical (age)
and/or numeric (length) values is to be performed.  In general:
\itemize{
\item Five age classes are simulated (if age is considered)
\item Mean and standard deviation of length increase with age (if length and age are considered)
\item Data are treated as belonging to four temporal strata (if stratified), in which:
\itemize{
\item Age class generally decreases with stratum (older fish arrive sooner)
\item Abundance and its standard error both increase with stratum (larger numbers of fish arrive later), but
\item An equal number of fish are sampled per temporal stratum.
}
}

It should be noted that a small but non-trival number of entries for age,
length, and stratum are imputed as \code{NA} by default (see \verb{nNA=} argument) in order
to test function robustness to missing data.
}
\examples{
# running using a pre-defined case
par(mfrow=c(2,2))
verify_ASL_table(case="stratified_witherror_lengthage", nsim=500)

# running using simulation parameters directly
par(mfrow=c(2,2))
verify_ASL_table(nstrata = 4,
                 nage = 5,

                 # sample size for each stratum
                 nt = c(100, 100, 100, 100),

                 # abundance for each stratum
                 Nt = c(10000, 20000, 30000, 40000),

                 # (possible) SE for abundance by stratum
                 se_Nt = 0.2*c(10000, 20000, 30000, 40000),

                 # mean length FOR EACH AGE
                 mn_length = c(150, 200, 250, 300, 350),

                 # sd length FOR EACH AGE
                 sd_length = c(40, 50, 60, 70, 80),

                 # matrix of probabilities of each age BY stratum
                 ptz = matrix(c(c(1,2,3,4,5),
                                c(1,2,5,5,2),
                                c(2,5,3,2,1),
                                c(5,4,3,1,1)), byrow=TRUE, nrow=4, ncol=5),
                 nsim=1000)

# simulating first, then running
par(mfrow=c(2,2))
simresults <- simulate_ASL_table(case="stratified_witherror_lengthage", nsim=500)
verify_ASL_table(simresults)


# all pre-defined cases
\dontrun{
nsim <- 5000
cases <- c("stratified_witherror_lengthage", "stratified_witherror_age",
"stratified_witherror_length", "stratified_lengthage", "stratified_age",
"stratified_length", "stratified_Nunknown_lengthage", "stratified_Nunknown_age",
"stratified_Nunknown_length", "pooled_witherror_lengthage", "pooled_witherror_age",
"pooled_witherror_length", "pooled_lengthage", "pooled_age", "pooled_length",
"pooled_Nunknown_lengthage", "pooled_Nunknown_age", "pooled_Nunknown_length")
for(case_i in cases) {
  par(mfrow=c(3,2))
  verify_ASL_table(case=case_i, nsim=nsim)
}
}
}
\seealso{
\link{ASL_table}
}
\author{
Matt Tyers
}
